feat: RESTful-show-response (#63)
* feat: RESTful-show-response

* refactor: change restful form placement

* style: refactor styles

* refactor: remove logging
