+++ package.json
@@ -24,6 +24,7 @@
     "react": "^18",
     "react-dom": "^18",
     "react-hook-form": "^7.53.0",
+    "react-json-formatter": "^0.4.0",
     "yup": "^1.4.0"
   },
   "devDependencies": {
