+++ src/api/get-restful-data.tsx
@@ -1,4 +1,4 @@
-import { SearchParams } from '@/types';
+import { ApiResponse, SearchParams } from '@/types';
 import { MethodType } from '@/types/restful';
 
 export default async function getRestfulData(
@@ -6,15 +6,19 @@ export default async function getRestfulData(
   url: string,
   searchParams?: SearchParams,
   body?: string,
-) {
-  console.log('body: ', body);
+): Promise<ApiResponse> {
+  const response: ApiResponse = { data: undefined, status: { code: undefined, text: undefined } };
   const options = body ? { method: method, headers: searchParams, body } : { method: method, headers: searchParams };
-  console.log('options: ', options);
-  const data: unknown = await fetch(url, options)
-    .then((res) => res.json())
-    .catch((error: Error) => {
-      console.log(error);
-    });
+  await fetch(url, options)
+    .then((res: Response) => {
+      response.status.code = res.status;
+      response.status.text = res.statusText;
+      return res.json();
+    })
+    .then((data) => {
+      response.data = data;
+    })
+    .catch((error: Error) => error);
 
-  return data;
+  return response;
 }
