+++ src/app/[lang]/globals.scss
@@ -43,7 +43,6 @@ footer {
 
 main {
   flex-grow: 1;
-  height: 1000px; // to test sticky header
 }
 
 button {
