+++ src/components/response-viewer/response-viewer.module.scss
@@ -12,6 +12,8 @@
   display: flex;
   flex-direction: column;
   gap: 10px;
+  max-height: 400px;
+  overflow-y: auto;
 }
 
 .status-code {
