+++ src/components/restful-client/restful.module.scss
@@ -1,0 +1,35 @@
+.select {
+  width: 120px;
+}
+
+.inputUrl {
+  flex-grow: 1;
+}
+
+.button {
+  font-size: 1rem;
+}
+
+.formInputLine {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: flex-end;
+  gap: 5px;
+  width: 100%;
+  margin-top: 15px;
+}
+
+.fieldset {
+  padding: 5px;
+  border: 1px solid var(--table-border);
+
+  .sectionTitle {
+    font-size: 1.1rem;
+    padding: 0 3px;
+    font-family: var(--font-family-mui);
+    font-weight: 400;
+    line-height: 1.5em;
+    color: var(--form-legend-color);
+  }
+}
